pyproject.toml
 [tool.poetry]
 name = "gcp-speech-to-speech-translation"
 version = "0.1.0"
 description = "Live vertaalservice met TDD"
 authors = ["Jonathan Yan <jonathanyan@example.com>"]
 readme = "README.md"
 packages = [{include = "backend"}]

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

 [tool.poetry.dependencies]
 python = "^3.9"
 fastapi = "^0.109.2"
 uvicorn = {extras = ["standard"], version = "^0.27.0"}
 websockets = "^12.0"
 tenacity = "^9.1.2"
 pydantic-settings = "^2.10.1"
 pybreaker = "1.4.0"
 google-cloud-speech = "^2.21.0"
 google-cloud-translate = "^3.21.1"
 google-cloud-texttospeech = "^2.27.0"
 pydub = "^0.25.1"
 psutil = "^7.0.0"
 google-cloud-logging = "^3.10.0"
 google-cloud-monitoring = "^2.21.0"
 prometheus-client = "^0.19.0"

 [tool.poetry.group.dev.dependencies]
 pytest = "^8.0.0"
 httpx = "^0.26.0"
 pytest-asyncio = "^0.23.5"

 [build-system]
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"